Reorganize functional test suite
Organize generic test suite into classes
Organize project specific test suite into classes
